Implementation change: now the dynamic client generates the client classes in a background thread as soon as the client URL is set. The init() method of course waits for the initialisation to finish if needed, checks the client is correctly initialised and connects.
Users of the dynamic client are advised to call setIcatBaseURL() as soon
as possible in their code to let the client generate its classes while
they perform other actions.

We kept the spring convention that init should be called only once, but
with this implement the setIcatBaseURL method can only be called once
too. This simplifies the implementation while not having too much impact
on users, all current use case have one client per url and instantiate a
new client for new urls.